# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ETMS API
  version: "1.0.0"
  description: >
    Employee Transportation Management System REST API. Human-readable contract
    in docs/etms/06-api-spec.md; this file covers the endpoints implemented in the
    backend service. All /v1 routes require a Bearer JWT carrying tenant + permissions.
servers:
  - url: https://api.{region}.etms.app/v1
    variables: { region: { default: dev } }
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        request_id: { type: string }
    LatLng:
      type: object
      required: [lat, lng]
      properties: { lat: { type: number }, lng: { type: number } }
security:
  - bearerAuth: []
paths:
...
  /v1/routes/optimize:
    post:
      summary: Route Optimization Engine — capacitated nearest-neighbour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [depot, capacity, stops]
              properties:
                depot: { $ref: "#/components/schemas/LatLng" }
                capacity: { type: integer, minimum: 1 }
                stops:
                  type: array
                  items:
                    allOf:
                      - $ref: "#/components/schemas/LatLng"
                      - type: object
                        required: [id, demand]
                        properties:
                          id: { type: string }
                          demand: { type: integer, minimum: 0 }
      responses:
        "200": { description: Optimized routes }
        "403": { description: Forbidden, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
  /v1/checkin/verify:
    post:
      summary: Verify a QR token and record a trip event (idempotent)
      responses:
        "200": { description: Verified }
        "409": { description: Invalid/expired QR }
...
  /v1/reports/operational:
    get:
      summary: Operational report (trips/on-time/no-shows) for a date range
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
      responses: { "200": { description: Aggregated metrics } }